Findbugs: Testcase and comment for false positive
Findbugs reports what appears to be a legitimate error that
is actually caught by logic long-preceeding this codepath.
Add a test case since that seems fragile and a few lines of
code to satisfy findbugs, and inform future readers of what
is happening ... since neither izzy nor I predicted the actual
result.